Rework app structure, separate Task Details screen

+++ domain/src/main/java/szysz3/planty/domain/usecase/task/GetTaskByIdUseCase.kt
@@ -7,8 +7,8 @@ import javax.inject.Inject
 
 class GetTaskByIdUseCase @Inject constructor(
     private val repository: TaskRepository
-) : BaseUseCase<Int, Task?>() {
-    override suspend fun invoke(input: Int): Task? {
-        return repository.getTaskById(input.toLong())
+) : BaseUseCase<Long, Task?>() {
+    override suspend fun invoke(input: Long): Task? {
+        return repository.getTaskById(input)
     }
 }
